EventManager.cs & CrackerParticleEffectHandler.cs Updated & BoardManager.cs Created
EventManager.cs is the only static class that holds data in the project so It was not a good idea to put the OnApplicationQuit method to StaticInstance abstract class to clear the eventHandlers for being able to disable domain reload. So I removed it from StaticInstance class and put it into EventManager.cs.