ItemRemovalType enum created and InGameUI Prefab updated
It is not possible to choose which item to explode in a cell and I don't want to do any changes to the current version since this project took more time than it needed. So I just created an enum in order to make the removal logic work with different item types.