Global Volume Prefab Updated And PostProcessingManager.cs is created
Bloom is really resource intensive so I made a safety net to use when needed. I don't know how will be the performance of my game in a real device so I'm not using the bloom opening and closing events. But it works pretty well based on my tests aside from constantly invoking an event for it.