EventManager.cs Optimized and A bug about Removing Handlers is fixed
There was a bug that made the last element of a handler list impossible to remove. It gave me a lot of problems since EventManager was persistent and it was blocking my classes from getting collected by GC. I handled it by simply checking if list count is 1 and if it is 1 I'm just clearing the whole list.