MainEventPanel and MainEventGoalObtainedPanel classes are created
Namings are little bit confusing so here is some clarification. MainEventPanel Handles all the main event logic while MainEventGoalObtainedPanel is just for visual animations for obtained goals in a level. It can also be used as a generic class for all kinds of obtained items such as cards, boosters etc.