LightBallBooster.cs and LightBall Merge actions are updated
Now it fires events with position itemID and idToMatch. I'm finding the most common item if it gets exploded instead of swapped but I'm not sure that if its the best thing to do. Also it understands the difference between merge action and Normal Action by just checking if its id is lower than 100 or not(Not very generic but I couldn't find a simpler solution)