Quit Level Panel Prefab And Try Again Panel Prefab are created
Quit level panel has a broken hard and A "You will use a heart" text.(Which doesn't really do anything but shhhh its between us). Also Try Again Panel is just a sibling of Level Panel with a Level Failed Text and Goals That Aren't completed panel Addition on top of it.