On Top Of Board Canvas prefab is created and Obstacle Particle Effect prefabs are updated
On Top Of Board Canvas is came out of necessity since I couldn't find any solution to fade every UI element except a particular element. So I divided the canvas into two with putting power up panels to front and in game UI to under it so that I can use 2 fades : one in front canvas and one in under canvas. This enables me to make board and elements in it unfaded while making power up buttons and move-goal panel faded.(Its hard to explain)