Settings Canvas Prefab is created And Boosters And Settings Panel prefab is updated
Old version of Boosters And Settings Panel was not responsive for different resolutions and It was not able to make fade everything except the clicked button. I added a fade image and I'm changing its order and clicked button's order in hierarchy so that everything except clicked button is getting faded.