LightBallBoosterAction.cs and LightBallRay.cs's Animate Ray method are updated
There was a check that is trying to understand if the item that is getting highlighted by ray is matching or exploding. It was causing too many issues at runtime such as making wrong items locked or unlocked or adding items to ItemsToExplode list without highlighting them. In order to solve the issue I removed the check and just made the item not active and locked.