Settings Canvas Updated and Slider Settings Button  Prefab is created
At previous version I needed to make changes on all 4 of the settings slider buttons even if they are exactly the same. So I created a prefab and added it instead of non-prefab buttons in settings page. Also I made some adjustments so that it fits into different resolutions as well